requirements.txt
black==24.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
moviepy==1.0.3
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
proglog==0.1.10
python-slugify==8.0.2
pytube==15.0.0
requests==2.31.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
